// src/Data/Models/SignUpModel.ts
export type SignUpBaseData = {
  first_name: string;
  second_name: string;
  first_lastname: string;
  second_lastname: string;
  email: string;
  password: string;
  address: string;
  cell_phone: string;
  dni: string,
  birth_date: string;
}

export type SignUpBase = {
  first_name: string;
  second_name: string;
  first_lastname: string;
  second_lastname: string;
  email: string;
  password: string;
}